made code modular by creating a seperate file for name menu
