+++ supa-fighta-client/src/game/stateManager.py
@@ -2,6 +2,7 @@ from game.states.mainMenu import MainMenuState
 from game.states.gameplayState import GameplayState
 from game.states.lobby import LobbyState
 from game.states.settingsMenu import SettingsState
+from game.states.nameMenu import NameMenuState
 from animations.sprites import SpriteSheet
 from animations.animation import Animator
 from type.sprite import SpriteProperties
@@ -9,8 +10,7 @@ import config
 
 class GameState:
     def __init__(self):
-        self.current_state = None
-        self.previous_state = None
+        self.state_stack = []
         self.background_sprites = SpriteSheet(
             SpriteProperties(
                 path="assets/background.png",
@@ -25,43 +25,46 @@ class GameState:
         self.states = {
             "main_menu": MainMenuState(self),
             "lobby": LobbyState(self),
-            "settings": SettingsState(self)
+            "settings": SettingsState(self),
+            "name_menu": NameMenuState(self)
         }
         self.change_state("main_menu")
 
     def change_state(self, new_state: str):
-        print(f"Changed state to: {new_state}")
+        if config.DEBUG:
+            print(f"Changed state to: {new_state}")
         if new_state == "gameplay":
             lobby_state = self.states["lobby"]
             self.states["gameplay"] = GameplayState(lobby_state.get_player(), self)
-        
-        if self.current_state:
-            self.previous_state = self.current_state
-            self.current_state.exit()
-            self.current_state = None
-        self.current_state = self.states.get(new_state)
-        if self.current_state:
-            self.current_state.enter()
+        # reset stack so we don't have to worry about going back to old states with old data
+        self.state_stack = []
+        state = self.states.get(new_state)
+        if state:
+            self.state_stack.append(state)
+            state.enter()
 
-    def go_back(self):
-        if self.previous_state:
-            # Find the state name by comparing objects
-            for state_name, state_obj in self.states.items():
-                if state_obj is self.previous_state:
-                    self.change_state(state_name)
-                    break
+    def push_state(self, state):
+        self.state_stack.append(state)
+
+    def pop_state(self):
+        if len(self.state_stack) > 1:
+            self.state_stack.pop()
+
+    def current_state(self):
+        return self.state_stack[-1] if self.state_stack else None
 
     def update(self):
-        if self.current_state:
-            self.current_state.update()
-            
+        for state in self.state_stack:
+            state.update()
+
     def draw(self, screen):
-        if self.current_state:
-            self.current_state.draw(screen)
-            
+        for state in self.state_stack:
+            state.draw(screen)
+
     def handle_event(self, event):
-        if self.current_state:
-            self.current_state.handle_event(event)
+        if self.state_stack:
+            # only top gets input
+            self.state_stack[-1].handle_event(event)
 
     def draw_background(self, screen):
         self.background.draw(screen)
